fix: inject reviewer trace context as JSON
